refactor

+++ LFG_Services/src/main/java/hu/blog/megosztanam/rest/PostRestController.java
@@ -1,9 +1,11 @@
 package hu.blog.megosztanam.rest;
 
+import hu.blog.megosztanam.model.AuthenticatedUser;
 import hu.blog.megosztanam.model.shared.GameMap;
 import hu.blog.megosztanam.model.shared.Post;
 import hu.blog.megosztanam.model.shared.summoner.Server;
 import hu.blog.megosztanam.service.IPostService;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.constraints.NotNull;
@@ -28,14 +30,16 @@ public class PostRestController {
     public List<Post> searchPosts(@PathVariable Server server,
                                   @RequestParam("userId") Integer userId,
                                   @RequestParam(name = "map", required = false) GameMap map,
-                                  @RequestParam(name = "isRanked",required = false) Boolean isRanked
+                                  @RequestParam(name = "isRanked", required = false) Boolean isRanked
     ) {
         return postService.getSearchForMemberPosts(server, userId, map, isRanked);
     }
 
     @RequestMapping(value = "{userId}/posts/{postId}", method = RequestMethod.DELETE)
-    public void deletePost(@PathVariable("userId") Integer userId, @PathVariable("postId") Integer postId){
-        postService.deletePost(userId, postId);
+    public void deletePost(@AuthenticationPrincipal AuthenticatedUser user, @PathVariable("userId") Integer userId, @PathVariable("postId") Integer postId) {
+        if(user.getUserId() == userId) {
+            postService.deletePost(userId, postId);
+        }
     }
 
 }
